Remove legacy termios
