+++ core/native/src/main/scala/terminus/NativeTerminal.scala
@@ -28,8 +28,8 @@ import scalanative.unsafe.*
 /** A Terminal implementation for Scala Native. */
 object NativeTerminal extends Terminal, WithEffect[Terminal] {
   val termios =
-    if LinktimeInfo.isMac then UnifiedTermios[TermiosTypes.mac]
-    else if LinktimeInfo.isLinux then UnifiedTermios[TermiosTypes.linux]
+    if LinktimeInfo.isMac then Termios[TermiosTypes.mac]
+    else if LinktimeInfo.isLinux then Termios[TermiosTypes.linux]
     else
       sys.error(
         s"""Your platform, {LinktimeInfo.target.os}, is not currently supported by Terminus on Scala Native.
