--- core/native/src/main/scala/terminus/UnifiedTermios.scala
@@ -1,33 +1,0 @@
-/*
- * Copyright 2024 Creative Scala
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package terminus
-
-import scala.scalanative.posix
-import scala.scalanative.unsafe.{Ptr, Zone}
-
-class UnifiedTermios[T](using accessor: TermiosAccess[T]) {
-  def getAttributes()(using Zone): Ptr[T] = accessor.get
-
-  def setAttributes(attributes: Ptr[T]): Unit = accessor.set(attributes)
-
-  def setRawMode(): Unit =
-    Zone {
-      val attrs = accessor.get
-      attrs.removeLocalFlags(posix.termios.ECHO | posix.termios.ICANON)
-      accessor.set(attrs)
-    }
-}
